--- Client/src/Components/BackgroundPattern/BackgroundPattern.jsx
@@ -1,9 +1,0 @@
-import "./backgroundPattern.css";
-import React from "react";
-import Pattern from "../../assets/Images/background_pattern_decorative.png";
-
-const BackgroundPattern = () => {
-  return <img className="background-pattern" src={Pattern} alt="Pattern" />;
-};
-
-export default BackgroundPattern;
