--- Client/src/Components/BackgroundPattern/backgroundPattern.css
@@ -1,7 +1,0 @@
-.background-pattern {
-  position: absolute;
-  top: -5%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  z-index: -1;
-}
