+++ Client/src/Components/Dashboard/DashboardMenu/index.jsx
@@ -1,15 +1,15 @@
 import "./index.css";
 import DashboardMenuButton from "../DashboardMenuButton";
 
-import Monitors from "../../assets/Images/Icon-monitor-gray.png";
-import Incidents from "../../assets/Images/Icon-warning-gray.png";
-import SensorsIcon from "../../assets/Images/Icon-signal-gray.png";
-import AllInclusiveIcon from "../../assets/Images/Icon-link-gray.png";
-import MaintenanceIcon from "../../assets/Images/Icon-maintenance-gray.png";
-import SettingsIcon from "../../assets/Images/Icon-setting-gray.png";
+import Monitors from "../../../assets/Images/Icon-monitor-gray.png";
+import Incidents from "../../../assets/Images/Icon-warning-gray.png";
+import SensorsIcon from "../../../assets/Images/Icon-signal-gray.png";
+import AllInclusiveIcon from "../../../assets/Images/Icon-link-gray.png";
+import MaintenanceIcon from "../../../assets/Images/Icon-maintenance-gray.png";
+import SettingsIcon from "../../../assets/Images/Icon-setting-gray.png";
 import React, { useState, useEffect } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
-import pathMap from "../../Utils/PathMap";
+import pathMap from "../../../Utils/PathMap";
 
 /**
  * @component
