--- Client/src/Components/Dropdown/index.css
@@ -1,1 +1,0 @@
-/* dropdown styles*/
