--- Client/src/Components/DropdownTeamMember/index.css
@@ -1,1 +1,0 @@
-/*Dropdown Team Member styles*/
