--- Client/src/Components/Icons/ComplexAlert/complexAlert.css
@@ -1,44 +1,0 @@
-.icon-holder-outer-red {
-  width: 38px;
-  height: 38px;
-  border-radius: 50%;
-  border: 2px solid #d92d2019;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.icon-holder-inner-red {
-  width: 28px;
-  height: 28px;
-  border-radius: 50%;
-  border: 2px solid #d92d204d;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.icon-holder-outer-green {
-  width: 38px;
-  height: 38px;
-  border-radius: 50%;
-  border: 2px solid #07945519;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.icon-holder-inner-green {
-  width: 28px;
-  height: 28px;
-  border-radius: 50%;
-  border: 2px solid #0794554d;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.icon-core {
-  width: 20px;
-  height: 20px;
-}
