--- Client/src/Components/MonitorTable/index.css
@@ -1,112 +1,0 @@
-/* .host-row {
-  display: flex;
-  align-items: center;
-}
-
-.host-row a {
-  width: var(--env-var-img-width-2);
-  height: var(--env-var-img-width-2);
-  color: var(--env-var-color-5);
-  margin-right: 10px;
-  margin-bottom: 2px;
-}
-.host-row a svg {
-  width: var(--env-var-font-size-large);
-  height: var(--env-var-font-size-large);
-}
-.host-row .host-percentage {
-  font-size: var(--env-var-font-size-small);
-}
-
-.host-name {
-  width: fit-content;
-  margin-right: 10px;
-  font-weight: 700;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  white-space: nowrap;
-}
-
-.monitors .MuiPaper-root:has(table.MuiTable-root) {
-  box-shadow: none;
-  border: solid 1px var(--env-var-color-16);
-  border-radius: var(--env-var-radius-1);
-}
-.monitors .MuiTableBody-root .MuiTableRow-root:last-child .MuiTableCell-root {
-  border: none;
-}
-.monitors .MuiTableHead-root,
-.monitors .MuiTableRow-root:hover {
-  background-color: var(--env-var-color-13);
-}
-.monitors .MuiTableHead-root .MuiTableCell-root,
-.monitors .MuiTableBody-root .MuiTableCell-root {
-  font-size: var(--env-var-font-size-medium);
-}
-.monitors .MuiTableHead-root .MuiTableCell-root {
-  padding: var(--env-var-spacing-1) var(--env-var-spacing-2);
-  font-weight: 500;
-  color: var(--env-var-color-2);
-}
-.monitors .MuiTableHead-root span {
-  display: inline-block;
-  height: 17px;
-  width: 20px;
-  overflow: hidden;
-  margin-bottom: -2px;
-  margin-left: 2px;
-}
-.monitors .MuiTableHead-root span svg {
-  width: 20px;
-  height: 20px;
-}
-.monitors .MuiTableBody-root .MuiTableCell-root {
-  color: var(--env-var-color-5);
-  padding: 6px var(--env-var-spacing-2);
-}
-
-.monitors .MuiPagination-root {
-  flex: 1;
-  margin-top: 35px;
-  border: solid 1px var(--env-var-color-16);
-  border-radius: var(--env-var-radius-1);
-  padding: var(--env-var-spacing-1-plus) var(--env-var-spacing-2);
-}
-.monitors .MuiPagination-root ul {
-  justify-content: center;
-}
-.monitors .MuiPagination-root ul li button {
-  font-size: var(--env-var-font-size-medium);
-  color: var(--env-var-color-5);
-  font-weight: 500;
-}
-.monitors .MuiPagination-root ul li:first-child {
-  margin-right: auto;
-}
-.monitors .MuiPagination-root ul li:last-child {
-  margin-left: auto;
-}
-.monitors .MuiPagination-root ul li:first-child button,
-.monitors .MuiPagination-root ul li:last-child button {
-  border: solid 1px var(--env-var-color-16);
-  border-radius: var(--env-var-radius-1);
-}
-.monitors .MuiPagination-root ul li:first-child button {
-  padding: 0 var(--env-var-spacing-1) 0 var(--env-var-spacing-1-plus);
-}
-.monitors .MuiPagination-root ul li:last-child button {
-  padding: 0 var(--env-var-spacing-1-plus) 0 var(--env-var-spacing-1);
-}
-.monitors .MuiPagination-root ul li:first-child button::after,
-.monitors .MuiPagination-root ul li:last-child button::before {
-  position: relative;
-  display: inline-block;
-}
-.monitors .MuiPagination-root ul li:first-child button::after {
-  content: "Previous";
-  margin-left: 15px;
-}
-.monitors .MuiPagination-root ul li:last-child button::before {
-  content: "Next";
-  margin-right: 15px;
-} */
