--- Client/src/Components/Pagination/index.css
@@ -1,33 +1,0 @@
-.pagination-holder {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-}
-
-.MuiPagination-root {
-  .MuiPagination-ul {
-    flex-wrap: nowrap;
-    li {
-      &:first-child {
-        flex-basis: 100%;
-        display: flex;
-        justify-content: flex-start;
-        align-items: center;
-        > button::after {
-          margin-left: 10px;
-          content: "previous";
-        }
-      }
-      &:last-child {
-        flex-basis: 100%;
-        display: flex;
-        justify-content: flex-end;
-        align-items: center;
-        > button::before {
-          margin-right: 10px;
-          content: "next";
-        }
-      }
-    }
-  }
-}
