--- Client/src/Components/PopupModals/DualButtonPopupModal/dualButtonPopupModal.css
@@ -1,52 +1,0 @@
-.dual-button-popup-modal-holder {
-  width: 380px;
-  height: 130px;
-  margin: 10px 20px;
-  padding: 30px;
-  box-shadow: 0px 8px 8px -4px rgba(16, 24, 40, 0.03),
-    0px 20px 24px -4px rgba(16, 24, 40, 0.08);
-  border: 1px solid var(--env-var-color-9);
-  border-radius: var(--env-var-radius-1);
-}
-
-.dual-button-popup-modal-subject {
-  color: var(--env-var-color-1);
-  font-weight: bolder;
-  font-size: 16px;
-}
-
-.dual-button-popup-modal-description {
-  color: var(--env-var-color-2);
-  font-size: var(--env-var-font-size-medium);
-  margin-top: 10px;
-}
-
-.dual-modal-spacing {
-  height: 40px;
-}
-
-.dual-button-popup-modal-controllers {
-  display: flex;
-  justify-content: end;
-}
-
-.blue-save-button {
-  width: 145px;
-  height: 35px;
-  padding: 5px 20px;
-  background-color: var(--env-var-color-3);
-  border: 1px solid var(--env-var-color-10);
-  border-radius: var(--env-var-radius-1);
-  margin: 5px 0;
-  margin-left: 20px;
-  color: var(--env-var-color-8);
-  cursor: pointer;
-  font-size: var(--env-var-font-size-medium);
-}
-
-.transparent-discard-button {
-  border: none;
-  background-color: transparent;
-  margin: 5px 10px;
-  font-size: var(--env-var-font-size-medium);
-}
