--- Client/src/Components/PopupModals/DualButtonPopupModalWithTextfields/dualButtonPopupModalWithTextfields.css
@@ -1,55 +1,0 @@
-.popup-modal-holder {
-  width: 480px;
-  font-family: var(--popup-font-family-0);
-  padding: 50px 40px;
-  border-radius: var(--env-var-radius-1);
-  margin: 10px 20px;
-  border: 1px solid var(--env-var-color-9);
-  box-shadow: 0px 8px 8px -4px rgba(16, 24, 40, 0.03),
-    0px 20px 24px -4px rgba(16, 24, 40, 0.08);
-}
-
-.popup-modal-header {
-  display: flex;
-  justify-content: space-between;
-  font-weight: bold;
-}
-
-.popup-modal-close {
-  cursor: pointer;
-}
-
-.spacing-height {
-  height: 30px;
-}
-
-.popup-modal-input-title {
-  margin-bottom: 15px;
-  color: var(--env-var-color-5);
-}
-
-.popup-modal-text-field {
-  width: 100%;
-  height: 35px;
-  border: 1px solid var(--env-var-color-4);
-  border-radius: var(--env-var-radius-1);
-}
-
-.popup-modal-controllers {
-  display: flex;
-  justify-content: end;
-}
-
-.white-cancel-button {
-  border: none;
-  background-color: transparent;
-  font-family: var(--popup-font-family-0);
-  margin: 5px 10px;
-  width: 145px;
-  height: 35px;
-  padding: 5px 20px;
-  border: 1px solid var(--env-var-color-4);
-  border-radius: var(--env-var-radius-1);
-  cursor: pointer;
-  font-size: var(--env-var-font-size-medium);
-}
